Moving out conftest from cli
